Fixed learning rate arg type

+++ bertft.py
@@ -107,7 +107,7 @@ parser.add_argument("--model_name", type=str, default="bert-base-uncased", help=
 parser.add_argument("--task_name", type=str, default="cola", help="")
 parser.add_argument("--max_length", type=int, default=512, help="")
 parser.add_argument("--batch_size", type=int, default=32, help="")
-parser.add_argument("--learning_rate", type=int, default=2e-5, help="")
+parser.add_argument("--learning_rate", type=float, default=2e-5, help="")
 parser.add_argument("--seed", type=int, default=5, help="")
 parser.add_argument(
     "--freeze_bert",
